+++ src/main/java/org/wedding/application/port/in/usecase/card/ModifyCardUseCase.java
@@ -1,4 +1,4 @@
-package org.wedding.application.port.in.usecase;
+package org.wedding.application.port.in.usecase.card;
 
 import org.wedding.adapter.in.web.dto.ModifyCardRequest;
