+++ src/main/java/org/wedding/application/service/CardService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.wedding.adapter.in.web.dto.CreateCardRequest;
 import org.wedding.adapter.in.web.dto.ModifyCardRequest;
-import org.wedding.application.port.in.usecase.CreateCardUseCase;
-import org.wedding.application.port.in.usecase.ModifyCardUseCase;
+import org.wedding.application.port.in.usecase.card.CreateCardUseCase;
+import org.wedding.application.port.in.usecase.card.ModifyCardUseCase;
 import org.wedding.application.port.out.repository.CardRepository;
 import org.wedding.domain.card.Card;
 import org.wedding.domain.card.exception.CardError;
